feature: nprogress implemented
