+++ package.json
@@ -13,6 +13,7 @@
     "mongodb": "^3.6.2",
     "next": "10.0.0",
     "next-images": "^1.6.2",
+    "nprogress": "^0.2.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-toastify": "^6.0.9",
@@ -22,6 +23,7 @@
   "devDependencies": {
     "@types/mongodb": "^3.5.32",
     "@types/node": "^14.14.6",
+    "@types/nprogress": "^0.2.0",
     "@types/react": "^16.9.55",
     "@types/styled-components": "^5.1.4",
     "@types/yup": "^0.29.9",
