+++ src/components/Footer/styles.ts
@@ -3,6 +3,8 @@ import styled from 'styled-components'
 export const Container = styled.div`
   width: 100vw;
   height: 5rem;
+  position: absolute;
+  /* margin-top: 90px; */
 
   display: flex;
   flex-direction: row;
